fix bug

+++ ffmpeg-concat-mp4-files.py
@@ -83,11 +83,13 @@ if __name__ == "__main__":
 	subprocess.call( cmd )
 
 	for f in mp4_tmp:
-		print "[Remove] mp4_tmp: ", f
-		os.remove(f)
+		if os.path.exists(f):
+			print "[Remove] mp4_tmp: ", f
+			os.remove(f)
 	for f in wget_tmp:
-		print "[Remove] wget_tmp: ", f
-		os.remove(f)
+		if os.path.exists(f):
+			print "[Remove] wget_tmp: ", f
+			os.remove(f)
 
 	print
 	print "result: ", args.output
